+++ src/main.rs
@@ -1,22 +1,36 @@
 use std::sync::Arc;
 
 use tokio::net::TcpListener;
+use tracing::info;
 
-use zumic::{server::handle_connection, InMemoryStore, StorageEngine};
+use zumic::{server::handle_connection, InMemoryStore, Settings, StorageEngine, StorageType};
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> anyhow::Result<()> {
+    let settings = Settings::load()?;
+    info!("Loaded config: {:#?}", &settings);
+
     #[allow(clippy::arc_with_non_send_sync)]
-    let engine = Arc::new(StorageEngine::Memory(InMemoryStore::new()));
-    let listener = TcpListener::bind("127.0.0.1:6379").await?;
-    println!("Listening on 127.0.0.1:6379");
+    let engine = match settings.storage_type {
+        StorageType::Memory => Arc::new(StorageEngine::Memory(InMemoryStore::new())),
+        StorageType::Persistent => {
+            unimplemented!("Persistent storage not yet implemented")
+        }
+        StorageType::Cluster => {
+            unimplemented!("Cluster storage not yet implemented")
+        }
+    };
+
+    let listener: TcpListener = TcpListener::bind(settings.listen_address).await?;
+    info!("Listening on {}", settings.listen_address);
 
     loop {
         let (socket, addr) = listener.accept().await?;
-        println!("Client: {addr}");
-        // Обрабатываем соединение **последовательно** в текущем потоке
-        if let Err(e) = handle_connection(socket, engine.clone()).await {
-            eprintln!("Connection error: {e:?}");
+        info!("Client connected: {}", addr);
+
+        let eng = engine.clone();
+        if let Err(e) = handle_connection(socket, eng).await {
+            tracing::error!("Connection error: {e:?}");
         }
     }
 }